Add "edit" logic and toolbar functions
Add edit logic in. The hard part will be re-generating the tree view with e appropriate types.

You can select from the combo box a new type for a given key and it will set it in the view, just need to hook that up to actually do so.

Say you had a key that was a dict, you should be able to change it to an Integer type and the dict will be "destroyed" and you can enter a new value for the new Integer type that was just created for it.

Going from not dict/array will be the toughest in representation. I will probably create an add / remove button to the toolbar that will add or remove a new key entry at the selected key. If the selected key is an array or dict it will be added to it. If the select type is not one of those two, it will go the the parent scope.