Add file opening
Add logic to open files with the menu bar: File->Open.

It will open the current directory of where the script is being ran from.

I also found a bug where all the nodes under the root are disabled for editing, but the children of those are still editable.

Another is that they can be "edited" while editing is disabled.