Add correct type checking on value change
Add the correct type of type checking when a value is changed for a node in the tree. Previously it used isinstance() which would only return true on string since the value field is a string (the text of the QStandardItem).

Changed this to check with both regex, datetime parsing, and boolean (True or False) strings.